Split shm; and numpy to extras
